requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
black==25.1.0
click==8.2.1
fastapi==0.116.1
flake8==7.3.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy==1.17.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
